Modularized gamelogic code.
