+++ server/collisions.js
@@ -1,3 +1,6 @@
+//collisions.js
+
+
 var Collisions = module.exports = {};
 
 Collisions.collisionDetection = function(player, collisionObject) {
@@ -7,7 +10,7 @@ Collisions.collisionDetection = function(player, collisionObject) {
   var distanceToFlag = Math.sqrt(distanceToFlagX + distanceToFlagY);
 
   if (distanceToFlag <= player.radius + collisionObject.radius) {
-    console.log('collided!');
+    //console.log('collided!');
     return true;
   }
   else return false;
@@ -49,7 +52,7 @@ Collisions.windowDetection = function(position, direction) {
       return true;
     }
     else {
-      console.log('we be stuck X');
+      //console.log('we be stuck X');
       return false;
     }
   }
@@ -58,7 +61,7 @@ Collisions.windowDetection = function(position, direction) {
       return true;
     }
     else {
-      console.log('we be stuck Y');
+      //console.log('we be stuck Y');
       return false;
     }
   }
