+++ server/score.js
@@ -1,0 +1,17 @@
+//this module is needed for events.js
+
+
+var handleWin = function() {
+
+};
+
+
+var checkForWin = function() {
+
+};
+
+
+module.exports.updateScore = function() {
+
+  //check for win
+};
